Updated references to TFConfig and moved TFProjectile out of TerrariaFortress.cs.
